requirements.txt
flask
python-dotenv
requests
discord.py
